modified Config.java and Buffer.java

+++ src/main/java/com/clevertap/stormdb/exceptions/ReservedKeyException.java
@@ -5,7 +5,7 @@ package com.clevertap.stormdb.exceptions;
  */
 public class ReservedKeyException extends RuntimeException {
 
-    public ReservedKeyException(final int key) {
-        super("The key 0x" + Integer.toHexString(key) + " is reserved for internal structures.");
+    public ReservedKeyException(final long key) {
+        super("The key 0x" + Long.toHexString(key) + " is reserved for internal structures.");
     }
 }
